servicios

+++ src/propuesta/propuesta.service.ts
@@ -21,13 +21,16 @@ export class PropuestaService {
         return propuesta;
     }
     async crearPropuesta(propuesta: PropuestaEntity): Promise<PropuestaEntity> {
+        if (propuesta.titulo.length==0)
+            throw new BusinessLogicException("The propuesta's title must not be empty'", BusinessError.NOT_FOUND);
         return await this.propuestaRepository.save(propuesta);
     }
     async deletePropuesta(id: number) {
-        const propuesta: PropuestaEntity = await this.propuestaRepository.findOne({where:{id}});
+        const propuesta: PropuestaEntity = await this.propuestaRepository.findOne({where:{id}, relations: ["proyecto"]});
         if (!propuesta)
           throw new BusinessLogicException("The propuesta with the given id was not found", BusinessError.NOT_FOUND);
-     
+        if (propuesta.proyecto != null)
+            throw new BusinessLogicException("The propuesta cannot be deleted if it has an active proyecto", BusinessError.PRECONDITION_FAILED);
         await this.propuestaRepository.remove(propuesta);
     }
 }
